Fix: Save exercises to hive from auth bloc instead of using exercise bloc
When the VerifyAuthenticationEvent is called from main.dart, await repositorioEjercicios.fetchExercises() is called.
Said function stores the exercises pulled from the api request in a hive box.

Furthermore, now the exercises shown on lista_ejercicios and further views are pulled from the hive box instead of using an api request.