Feat: Show exercices images in different views
Exercise images now show in details and item_ejercicios view.
This is made by doing a request to the API and getting the imageUrl of a specific exercise by its id.

This could be improved by storing the images in the device or in a remote DB instead of doing a network request of the image.