Feat: Functionality to add and remove an exercise to the workoutExercises List
Added a new boolean property in the Exercise model called selectedForWorkout. With this, when an exercise is selected in the add_workout, the property changes.
If the property is true, then the exercise is added to the list, otherwise its removed. This assures that only one instance of the exercise is copied into the list.