Feat: Save workout in CloudFirebase
Added a new event in the add_workout bloc to handle the upload of workouts to the firebase cloud.
Had to create a new function in the Exercise model called exerciseToJson that transform the object to a json.
Since firebase stores a Map<string,dynamic>, a map composed of an exerciseId and the json of the exercise is sent to the cloud