Feat: Login with a registered email and password
Added new event in login/bloc/login_event to handle the login with credentials, and updated the login_bloc to handle the event.
Added new functions in user_auth_provider to handle the logout, signIn and signOut of email users

Theres a bug with this implementation: When logging in with valid credentials are proportionated, the view doesnot change to Menu(), only after a hot reload.